# openapi.yml
openapi: 3.1.0
info:
  title: tgarchive
  description: ''
  license:
    name: ''
  version: 0.1.3
paths:
  /api/v1/chats:
    get:
      tags:
      - routes
      operationId: list_chats
      parameters:
      - name: tracked
        in: query
        description: Only return tracked (collected) chats.
        required: false
        schema:
          type: boolean
      - name: sort
        in: query
        description: Sort order; default is by chat ID.
        required: false
        schema:
          $ref: '#/components/schemas/ChatSortDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/refresh:
    post:
      tags:
      - routes
      operationId: refresh_chats
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}:
    get:
      tags:
      - routes
      operationId: get_chat
      parameters:
      - name: chat_id
        in: path
        description: Marked Telegram chat ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/media-policy:
    get:
      tags:
      - routes
      operationId: get_media_policy
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaPolicyDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags:
      - routes
      operationId: set_media_policy
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaPolicyDto'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaPolicyDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/media/downloads:
    post:
      tags:
      - routes
      operationId: backfill_media
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaDownloadDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/messages:
    get:
      tags:
      - routes
      operationId: list_chat_messages
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: chat_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: sender_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: post_author
        in: query
        description: Exact channel post signature.
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: before
        in: query
        required: false
        schema:
          type: string
      - name: after
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: include_deleted
        in: query
        description: Also return deleted messages (flagged with `is_deleted`/`deleted_at`).
        required: false
        schema:
          type: boolean
      - name: exclude_bots
        in: query
        description: Hide messages from known bots (`senders.is_bot=1`); unknown senders are kept.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePageDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/messages/{message_id}:
    get:
      tags:
      - routes
      operationId: get_message
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: message_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: include_deleted
        in: query
        description: Also return the message if it was deleted.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/messages/{message_id}/context:
    get:
      tags:
      - routes
      operationId: message_context
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: message_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: before
        in: query
        description: Messages older than the anchor (0-100, default 20).
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: after
        in: query
        description: Messages newer than the anchor (0-100, default 20).
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageContextDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Anchor message not found (or deleted without include_deleted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/messages/{message_id}/media:
    get:
      tags:
      - routes
      operationId: message_media
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: message_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaDownloadDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/senders:
    get:
      tags:
      - routes
      operationId: list_chat_senders
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        description: Maximum senders returned (1-1000, default 100).
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      responses:
        '200':
          description: Senders with non-deleted messages in the chat, most messages first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SenderSummaryDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/sync:
    post:
      tags:
      - routes
      operationId: sync_chat
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: refetch
        in: query
        description: |-
          Re-download the whole history and backfill NULL sender/post-author/forward/attachment
          metadata of archived messages (text, versions and deletions are untouched). Resumable.
        required: false
        schema:
          type: boolean
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Chat is not tracked (code chat_not_tracked) or a conflicting job is active (code conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/chats/{chat_id}/tracking:
    put:
      tags:
      - routes
      operationId: track_chat
      parameters:
      - name: chat_id
        in: path
        description: Marked Telegram chat ID
        required: true
        schema:
          type: integer
          format: int64
      - name: backfill
        in: query
        description: Also enqueue a history sync for the chat (needs the running collector). Reuses an active job instead of duplicating it.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Chat is now tracked (idempotent); with backfill=true also carries the history sync job id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackChatDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Storage unavailable, or backfill requested without a running collector / with a full queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags:
      - routes
      operationId: untrack_chat
      parameters:
      - name: chat_id
        in: path
        description: Marked Telegram chat ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Chat is no longer tracked; stored messages are kept (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/media/{media_id}:
    get:
      tags:
      - routes
      operationId: get_media
      parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDownloadDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/media/{media_id}/archive:
    post:
      tags:
      - routes
      operationId: archive_media
      parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDownloadDto'
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDownloadDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/media/{media_id}/content:
    get:
      tags:
      - routes
      operationId: media_content
      parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Verified local image bytes
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/media/{media_id}/retry:
    post:
      tags:
      - routes
      operationId: retry_media
      parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaDownloadDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/messages:
    get:
      tags:
      - routes
      operationId: list_messages
      parameters:
      - name: chat_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: sender_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: post_author
        in: query
        description: Exact channel post signature.
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: before
        in: query
        required: false
        schema:
          type: string
      - name: after
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: include_deleted
        in: query
        description: Also return deleted messages (flagged with `is_deleted`/`deleted_at`).
        required: false
        schema:
          type: boolean
      - name: exclude_bots
        in: query
        description: Hide messages from known bots (`senders.is_bot=1`); unknown senders are kept.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePageDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/messages/search:
    get:
      tags:
      - routes
      operationId: search_messages
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      - name: chat_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: sender_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: post_author
        in: query
        description: Exact channel post signature.
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: before
        in: query
        description: Cursor from the previous page's `next_cursor`.
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: 'Only with `sort=time`: cursor for the newer direction.'
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: include_deleted
        in: query
        description: Also return deleted messages (flagged with `is_deleted`/`deleted_at`).
        required: false
        schema:
          type: boolean
      - name: exclude_bots
        in: query
        description: Hide messages from known bots (`senders.is_bot=1`); unknown senders are kept.
        required: false
        schema:
          type: boolean
      - name: sort
        in: query
        description: Result order; default `relevance`.
        required: false
        schema:
          $ref: '#/components/schemas/SearchSortDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePageDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/status:
    get:
      tags:
      - routes
      operationId: status
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/sync:
    post:
      tags:
      - routes
      operationId: sync_all
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/sync/status:
    get:
      tags:
      - routes
      operationId: sync_status
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusDto'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ChatDto:
      type: object
      required:
      - id
      - kind
      - tracked
      properties:
        id:
          type: integer
          format: int64
        kind:
          $ref: '#/components/schemas/ChatKindDto'
        stats:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ChatStatsDto'
            description: Aggregate archive/sync stats; omitted by `POST /chats/refresh`.
        title:
          type:
          - string
          - 'null'
        tracked:
          type: boolean
          description: Whether the chat is opted in for collection.
        username:
          type:
          - string
          - 'null'
...
    ChatSortDto:
      type: string
      description: Sort order of `GET /chats`.
      enum:
      - title
      - last_message
      - message_count
...
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    MediaPolicyDto:
      type: object
      required:
      - auto_archive
      properties:
        auto_archive:
          type: boolean
...
    MessagePageDto:
      type: object
      required:
      - items
      - has_more
      properties:
        has_more:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/MessageDto'
        next_cursor:
          type:
          - string
          - 'null'
          description: Opaque cursor; pass it back as `before` to get the next page.
...